[Feat] 타이머 설정 #10
