+++ TrackUs-iOS/TrackUs-iOS/Common/Extension/String+.swift
@@ -1,0 +1,15 @@
+//
+//  String+.swift
+//  TrackUs-iOS
+//
+//  Created by 석기권 on 5/21/24.
+//
+
+import Foundation
+
+extension String {
+    var asNumber: Int {
+        guard let number = Int(self) else { return 0 }
+        return number
+    }
+}
